rebuild the metadata in a more straightforward way
this avoids having to (very slowly) load the author metadata each time, and this should be extensible to other additional data that might be useful, such as the merge key.

the changes to main() should be extensible to the update process as a whole.

it also provides a way of extending/overriding the author metadata used (useful when the author isn't interesting enough to justify a WikiData entry).

this means there will pretty much always be some metadata produced, so remove that test.