some test collections of various sizes
```
for size in small:120 medium:500 large:2300  # xlarge:10000
do
  tools/make-books.py --size "${size##*:}" --seed "${size##*:}" --output "benchmarks/data/collections/${size%%:*}/"
done
```

the sizes were chosen with `np.geomspace(120, 10_000, 4).astype(int)`.

the metadata files (which can be derived) have not been included as they can be derived easily enough. the empty and xlarge collections haven't been included either, since the former doesn't work and the latter is huge and possibly not all that useful right now.